Constructing-a-number_statement
Problem Statement:
Manipulating numbers is at the core of a programmer's job. To test how well you know their properties, you are asked to solve the following problem.
You are given n non-negative integers a1, a2, ..., an. You want to know whether it's possible to construct a new integer using all the digits of these numbers such that it would be divisible by 3. You can reorder the digits as you want. The resulting number can contain leading zeros.

For example, consider the numbers 50, 40, 90 from which you have to construct a new integer as described above. Numerous arrangements of digits are possible; but we have illustrated one below. see figure on mentioned link.

Function Description:
Complete the function canConstruct which takes an integer array as input and return "Yes" or "No" based on whether or not the required integer can be formed.

For details, refer
https://www.hackerrank.com/challenges/constructing-a-number/